Update All APIs

+++ src/utils/deleteRelatedDocumentsPlugin.js
@@ -4,13 +4,20 @@ export const deleteRelatedDocumentsPlugin = function (schema, options) {
     { document: true, query: true },
     async function (next) {
       try {
-        const filter = this.getFilter(); // Get the query filter
-        const documents = await this.model.find(filter); // Get documents that match the filter
+        const filter = this.getFilter();
+        const documents = await this.model.find(filter);
 
         if (options?.relatedModels) {
           for (const doc of documents) {
             for (const { model, field } of options.relatedModels) {
-              await model.deleteMany({ [field]: doc._id });
+              if (doc.deletedAt) {
+                await model.updateMany(
+                  { [field]: doc._id },
+                  { deletedAt: new Date() }
+                );
+              } else {
+                await model.deleteMany({ [field]: doc._id });
+              }
             }
           }
         }
